refactor: concentrated all the util functions, and remover unnecessary files
Concentrated
