--- packages/parcel-plugin-svelte/lib/parcel-adapter.js
@@ -1,9 +1,0 @@
-function parcelExport() {
-  try {
-    return require('parcel');
-  } catch (e) {
-    return require('parcel-bundler');
-  }
-}
-
-module.exports = parcelExport();
